Prepare structure and test for service
