+++ src/main/java/com/safetynet/safetynetalerts/service/PersonService.java
@@ -1,6 +1,8 @@
 package com.safetynet.safetynetalerts.service;
 
 import com.safetynet.safetynetalerts.model.Person;
+import com.safetynet.safetynetalerts.model.dto.ChildInfoDTO;
+import com.safetynet.safetynetalerts.model.dto.PersonFullInfoDTO;
 import com.safetynet.safetynetalerts.repository.impl.PersonRepositoryImpl;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -87,11 +89,22 @@ public class PersonService {
     }
 
     /**
-     * Get all phone number of inhabitant associated to the given station number.
-     * @param stationNumber concerned
-     * @return set of all phone number.
+     * Get all children at the specified address with other inhabitant.
+     * @param address .
+     * @return list of children
      */
-    public Set<String> getAllPhoneByStationNumber(final int stationNumber) {
+    public Set<ChildInfoDTO> getAllChildrenByAddress(final String address) {
+        //TODO
+        return null;
+    }
+
+    /**
+     * Get all information about a person.
+     * @param firstName .
+     * @param lastName .
+     * @return person's info
+     */
+    public PersonFullInfoDTO getAllInfoByFirstNameAndLastName(final String firstName, final String lastName) {
         //TODO
         return null;
     }
