+++ src/test/java/com/safetynet/safetynetalerts/web/controller/PersonControllerTest.java
@@ -126,25 +126,6 @@ class PersonControllerTest {
     }
 
     @Test
-    void getAllPhoneByStationNumberExistingTest() throws Exception {
-        //TODO
-        int stationNumber = 1;
-        //when().thenReturn();
-        mockMvc.perform(get("/phoneAlert?firestation={stationNumber}", stationNumber))
-                .andExpect(status().isOk());
-
-    }
-
-    @Test
-    void getAllPhoneByStationNumberUnknownTest() throws Exception {
-        //TODO
-        int stationNumber = 6;
-        //when().thenReturn();
-        mockMvc.perform(get("/phoneAlert?firestation={stationNumber}", stationNumber))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
     void getAllChildrenByAddressExistingTest() throws Exception {
         //TODO
         String address = "address";
